refactor: merge loon-ops into loon-client, decouple client from server
Move the ops orchestration module into loon-client as pub mod ops.
Server calls are replaced with the ServerTransport trait from loon-types,
breaking the compile-time dependency from client to server.

- loon-client depends only on loon-types (+ loon-sim, rusqlite, etc.)
- loon-cli creates a LocalTransport that calls loon-server in-process
- xtask does the same
- loon-ops crate removed from workspace

Final workspace: loon-types, loon-server, loon-client, loon-cli,
loon-sim, loon-testkit (+ xtask). 12 → 7 crates total.
