+++ crates/loon-cli/Cargo.toml
@@ -14,12 +14,13 @@ anyhow.workspace = true
 clap = { version = "4.5", features = ["derive"] }
 clap_complete = "4.5"
 clap_mangen = "0.2"
-loon-ops = { path = "../loon-ops" }
+loon-client = { path = "../loon-client" }
+loon-server = { path = "../loon-server" }
 loon-types = { path = "../loon-types" }
+thiserror.workspace = true
 toml.workspace = true
 
 [dev-dependencies]
-loon-objectstore = { path = "../loon-objectstore" }
 loon-server = { path = "../loon-server" }
 loon-testkit = { path = "../loon-testkit" }
 serde_json.workspace = true
