+++ crates/loon-client/Cargo.toml
@@ -6,7 +6,9 @@ license.workspace = true
 rust-version.workspace = true
 
 [dependencies]
+anyhow.workspace = true
 serde.workspace = true
+serde_yaml.workspace = true
 loon-types = { path = "../loon-types" }
 loon-server = { path = "../loon-server" }
 loon-sim = { path = "../loon-sim" }
@@ -14,7 +16,7 @@ rusqlite.workspace = true
 thiserror.workspace = true
 serde_json.workspace = true
 sha2.workspace = true
+toml.workspace = true
 
 [dev-dependencies]
 loon-testkit = { path = "../loon-testkit" }
-serde_yaml.workspace = true
