+++ crates/loon-client/src/conflict.rs
@@ -9,12 +9,11 @@ use crate::state_db::{
 };
 use crate::testing::{ClientExecutionHooks, NoopClientExecutionHooks};
 use crate::upload::{upload_small_file_from_path, UploadError, UploadedContent};
-use loon_server::objectstore::ObjectStoreError;
-use loon_server::objectstore::keys::{
+use loon_types::object_store_keys::{
     conflict_artifact, conflict_artifact_archive, conflict_artifact_archive_prefix,
     conflict_artifact_prefix,
 };
-use loon_server::objectstore::ObjectStore;
+use loon_types::{ObjectStore, ObjectStoreError};
 use loon_types::{
     deterministic_conflict_id, deterministic_subtree_conflict_id, ChangeSeq,
     ConflictArtifactArchiveEnvelope, ConflictArtifactEnvelope, ConflictArtifactKind,
