+++ crates/loon-client/src/download.rs
@@ -1,6 +1,5 @@
-use loon_server::objectstore::ObjectStoreError;
-use loon_server::objectstore::keys::{blob, content_manifest};
-use loon_server::objectstore::ObjectStore;
+use loon_types::object_store_keys::{blob, content_manifest};
+use loon_types::{ObjectStore, ObjectStoreError};
 use loon_types::{
     content_manifest_digest_sha256, decode_content_manifest_json, sha256_digest,
     ContentBlockDescriptor, ContentManifestCodecError, ContentManifestEnvelope, NamespaceId,
