+++ crates/loon-client/src/testing.rs
@@ -12,8 +12,7 @@ use crate::local_apply::LocalApplyError;
 use crate::state_db::{
     ClientFileId, ConflictArtifactArchiveRow, ConflictArtifactRow, SqliteStateDb,
 };
-use loon_server::objectstore::ObjectStoreError;
-use loon_server::objectstore::{ByteRange, ObjectMetadata, ObjectStore, PutMode};
+use loon_types::{ByteRange, ObjectMetadata, ObjectStore, ObjectStoreError, PutMode};
 use loon_sim::faults::{
     FaultPlan, InjectedClientFault, InjectedStoreErrorKind, InjectedStoreOperation,
 };
