--- crates/loon-ops/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "loon-ops"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-rust-version.workspace = true
-
-[dependencies]
-anyhow.workspace = true
-loon-client = { path = "../loon-client" }
-loon-server = { path = "../loon-server" }
-loon-types = { path = "../loon-types" }
-serde.workspace = true
-serde_json.workspace = true
-serde_yaml.workspace = true
-thiserror.workspace = true
-toml.workspace = true
-
-[dev-dependencies]
-loon-testkit = { path = "../loon-testkit" }
