+++ xtask/src/ops.rs
@@ -1,7 +1,7 @@
 use anyhow::Result;
 
 pub fn run(args: impl IntoIterator<Item = String>) -> Result<String> {
-    loon_ops::run_args(args)
+    loon_client::ops::run_args(args, crate::transport::make_transport, crate::transport::open_store)
 }
 
 #[cfg(test)]
